third_party/blink/web_tests/external/wpt/webstorage/set.window-expected.txt

This is a testharness.js-based test.
PASS Setting property for key 9 on localStorage
PASS Setting property with toString for key 9 on localStorage
FAIL Setting property for key 9 on localStorage with data property on prototype assert_equals: expected "proto" but got "value"
FAIL Setting property for key 9 on localStorage with data property on prototype and existing item assert_equals: expected "proto" but got "existing"
FAIL Setting property for key 9 on localStorage with accessor property on prototype assert_array_equals: lengths differ, expected 1 got 0
PASS Setting property for key x on localStorage
PASS Setting property with toString for key x on localStorage
FAIL Setting property for key x on localStorage with data property on prototype assert_equals: expected "proto" but got "value"
FAIL Setting property for key x on localStorage with data property on prototype and existing item assert_equals: expected "proto" but got "value"
FAIL Setting property for key x on localStorage with accessor property on prototype assert_equals: expected "proto getter" but got "value"
PASS Setting property for key 9 on sessionStorage
PASS Setting property with toString for key 9 on sessionStorage
FAIL Setting property for key 9 on sessionStorage with data property on prototype assert_equals: expected "proto" but got "value"
FAIL Setting property for key 9 on sessionStorage with data property on prototype and existing item assert_equals: expected "proto" but got "existing"
FAIL Setting property for key 9 on sessionStorage with accessor property on prototype assert_array_equals: lengths differ, expected 1 got 0
PASS Setting property for key x on sessionStorage
PASS Setting property with toString for key x on sessionStorage
FAIL Setting property for key x on sessionStorage with data property on prototype assert_equals: expected "proto" but got "value"
FAIL Setting property for key x on sessionStorage with data property on prototype and existing item assert_equals: expected "proto" but got "value"
FAIL Setting property for key x on sessionStorage with accessor property on prototype assert_equals: expected "proto getter" but got "value"
Harness: the test ran to completion.
